auto-email/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai>=0.7.0
protobuf>=4.25.0
python-dotenv==1.0.0
requests>=2.31.0
streamlit>=1.28.0